backend/pyproject.toml
[tool.poetry]
name = "ai4esg"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu118"
priority = "supplemental"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = "^0.109.0"
uvicorn = ">=0.20.0"
beautifulsoup4 = "^4.11.2"
openai = "^1.1.0"
requests = "^2.28.2"
numpy = "^1.24.2"
black = "^23.3.0"
transformers = "^4.37.0"
einops = "^0.6.1"
accelerate = "^0.26.1"
langchain = "^0.1.3"
pydantic = {version = "^2.0", extras = ["email"]}
safetensors = "^0.3.1"
pre-commit = "^3.3.3"
python-multipart = "^0.0.6"
torch = {version = "2.0.1+cu118", source = "pytorch"}
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
lxml = "^4.9.2"
sqlalchemy = "^2.0.17"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
alembic = "^1.11.1"
fastapi-sqlalchemy = "^0.2.1"
azure-storage-blob = "^12.16.0"
PyMuPDF = "^1.16.14"
websockets = "^11.0.3"
replicate = "^0.11.0"
jupyter = "^1.0.0"
tiktoken = "^0.3.3"
torchmetrics = "^1.2.1"
pypdf = "^4.0.0"
llama-index = "^0.9.40"
trulens-eval = "^0.20.0"
unstructured = {version = "^0.12.0", extras = ["pdf"]}
pgvector = "^0.1.4"

[tool.poetry.dev-dependencies]
optimum = "^1.16.2"
auto-gptq = "0.8.0.dev0+cu118"
llama-cpp-python = {version="0.2.44", extras = ["server"]}
gekko = "^1.0.6"
peft = "^0.8.2"
ninja = "^1.11.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
